feat: encapsulated spherical coordinates into histogram

+++ cpp/deplex/src/deplex/histogram.h
@@ -12,7 +12,7 @@
 namespace deplex {
 class Histogram {
  public:
-  Histogram(int32_t nr_bins_per_coord, Eigen::MatrixXd const& normals, std::bitset<BITSET_SIZE> const& mask);
+  Histogram(int32_t nr_bins_per_coord, Eigen::MatrixXf const& normals, std::bitset<BITSET_SIZE> const& mask);
 
   std::vector<int32_t> getPointsFromMostFrequentBin() const;
